Finished work, but lots of validation issues open
