+++ src/wn-adv.all.xml
@@ -26400,14 +26400,6 @@
       <Lemma writtenForm="derp" partOfSpeech="r"/>
       <Sense id="ewn-derp-r-90003183-01" n="0" synset="ewn-90003183-r" dc:identifier="derp%4:02:01::"/>
     </LexicalEntry>
-    <LexicalEntry id="ewn-glowy-a">
-      <Lemma writtenForm="glowy" partOfSpeech="a"/>
-      <Sense id="ewn-glowy-a-90019731-01" n="0" synset="ewn-90019731-a" dc:identifier="glowy%3:02:01::"/>
-    </LexicalEntry>
-    <LexicalEntry id="ewn-long_ass-a">
-      <Lemma writtenForm="long ass" partOfSpeech="a"/>
-      <Sense id="ewn-long_ass-a-90013262-01" n="0" synset="ewn-90013262-a" dc:identifier="long_ass%3:02:01::"/>
-    </LexicalEntry>
     <LexicalEntry id="ewn-so_yeah-r">
       <Lemma writtenForm="so yeah" partOfSpeech="r"/>
       <Sense id="ewn-so_yeah-r-91000331-01" n="0" synset="ewn-91000331-r" dc:identifier="so_yeah%4:02:01::"/>
@@ -45216,14 +45208,6 @@
     </Synset>
     <Synset id="ewn-90003183-r" ili="in" partOfSpeech="r" dc:subject="adv.all" dc:source="Colloquial WordNet">
       <Definition>reacting to a stupid but cute action</Definition>
-    </Synset>
-    <Synset id="ewn-90019731-a" ili="in" partOfSpeech="a" dc:subject="adv.all" dc:source="Colloquial WordNet">
-      <Definition>appearing to be glowing</Definition>
-      <SynsetRelation relType="similar" target="ewn-00077434-r"/> <!-- brightly, bright, brilliantly -->
-    </Synset>
-    <Synset id="ewn-90013262-a" ili="in" partOfSpeech="a" dc:subject="adv.all" dc:source="Colloquial WordNet">
-      <Definition>Very long in the temporal sense</Definition>
-      <SynsetRelation relType="hypernym" target="ewn-00167240-r"/> <!-- long -->
     </Synset>
     <Synset id="ewn-91000331-r" ili="in" partOfSpeech="r" dc:subject="adv.all" dc:source="Colloquial WordNet">
       <Definition>In other words, as a result</Definition>
