+++ src/wn-noun.Tops.xml
@@ -1184,7 +1184,6 @@
       <SynsetRelation relType="hyponym" target="ewn-87779642-n"/>
       <SynsetRelation relType="hyponym" target="ewn-90011301-n"/>
       <SynsetRelation relType="hyponym" target="ewn-90016671-n"/>
-      <SynsetRelation relType="hyponym" target="ewn-90005851-n"/>
       <SynsetRelation relType="hyponym" target="ewn-90004881-n"/>
       <SynsetRelation relType="hyponym" target="ewn-90009072-n"/>
       <Example>&quot;there was too much for one person to do&quot;</Example>
