+++ src/wn-noun.cognition.xml
@@ -46066,9 +46066,9 @@
       <SynsetRelation relType="hypernym" target="ewn-05652767-n"/> <!-- inability -->
     </Synset>
     <Synset id="ewn-90013991-n" ili="in" partOfSpeech="n" dc:subject="noun.cognition" dc:source="Colloquial WordNet">
-      <Definition>To correct automatically (of a computer system)</Definition>
-      <SynsetRelation relType="hypernym" target="ewn-06138021-n"/> <!-- computer science, computing -->
-      <SynsetRelation relType="hypernym" target="ewn-00200023-v"/>
+      <Definition>a system for automatically correcting text</Definition>
+      <SynsetRelation relType="domain_topic" target="ewn-06138021-n"/> <!-- computer science, computing -->
+      <SynsetRelation relType="hypernym" target="ewn-06581154-n"/>
     </Synset>
     <Synset id="ewn-90018351-n" ili="in" partOfSpeech="n" dc:subject="noun.cognition" dc:source="Colloquial WordNet">
       <Definition>something that inspires you to exercise</Definition>
