+++ src/wn-noun.food.xml
@@ -15190,9 +15190,9 @@
       <Lemma writtenForm="superfood" partOfSpeech="n"/>
       <Sense id="ewn-superfood-n-90017071-01" n="0" synset="ewn-90017071-n" dc:identifier="superfood%1:13:01::"/>
     </LexicalEntry>
-    <LexicalEntry id="ewn-takoyaki-v">
-      <Lemma writtenForm="takoyaki" partOfSpeech="v"/>
-      <Sense id="ewn-takoyaki-v-90011371-01" n="0" synset="ewn-90011371-v" dc:identifier="takoyaki%2:13:01::"/>
+    <LexicalEntry id="ewn-takoyaki-n">
+      <Lemma writtenForm="takoyaki" partOfSpeech="n"/>
+      <Sense id="ewn-takoyaki-n-90011371-01" n="0" synset="ewn-90011371-n" dc:identifier="takoyaki%2:13:01::"/>
     </LexicalEntry>
     <LexicalEntry id="ewn-white_tea-n">
       <Lemma writtenForm="white tea" partOfSpeech="n"/>
@@ -15597,7 +15597,7 @@
       <SynsetRelation relType="hyponym" target="ewn-07954667-n"/> <!-- slop -->
       <SynsetRelation relType="hyponym" target="ewn-90014051-n"/>
       <SynsetRelation relType="hyponym" target="ewn-90017071-n"/>
-      <SynsetRelation relType="hyponym" target="ewn-90011371-v"/>
+      <SynsetRelation relType="hyponym" target="ewn-90011371-n"/>
       <Example>&quot;food and drink&quot;</Example>
     </Synset>
     <!-- comfort food -->
@@ -32348,7 +32348,7 @@
       <Definition>a food with health benefits</Definition>
       <SynsetRelation relType="hypernym" target="ewn-07571428-n"/> <!-- food, solid food -->
     </Synset>
-    <Synset id="ewn-90011371-v" ili="in" partOfSpeech="v" dc:subject="noun.food" dc:source="Colloquial WordNet">
+    <Synset id="ewn-90011371-n" ili="in" partOfSpeech="n" dc:subject="noun.food" dc:source="Colloquial WordNet">
       <Definition>a ball-shaped Japanese snack made of a wheat flour-based batter and cooked in a special moulded pan</Definition>
       <SynsetRelation relType="hypernym" target="ewn-07571428-n"/> <!-- food, solid food -->
     </Synset>
