+++ src/wn-noun.time.xml
@@ -7526,9 +7526,9 @@
       <Lemma writtenForm="mornight" partOfSpeech="n"/>
       <Sense id="ewn-mornight-n-90009411-01" n="0" synset="ewn-90009411-n" dc:identifier="mornight%1:28:01::"/>
     </LexicalEntry>
-    <LexicalEntry id="ewn-pre-valentines_day-a">
-      <Lemma writtenForm="pre-valentines day" partOfSpeech="a"/>
-      <Sense id="ewn-pre-valentines_day-a-90001211-01" n="0" synset="ewn-90001211-a" dc:identifier="pre-valentines_day%3:28:01::"/>
+    <LexicalEntry id="ewn-pre-valentines_day-n">
+      <Lemma writtenForm="pre-valentines day" partOfSpeech="n"/>
+      <Sense id="ewn-pre-valentines_day-n-90001211-01" n="0" synset="ewn-90001211-n" dc:identifier="pre-valentines_day%3:28:01::"/>
     </LexicalEntry>
     <LexicalEntry id="ewn-snow_day-n">
       <Lemma writtenForm="snow day" partOfSpeech="n"/>
@@ -9435,7 +9435,7 @@
       <SynsetRelation relType="hyponym" target="ewn-15247636-n"/> <!-- washing day, washday -->
       <SynsetRelation relType="hyponym" target="ewn-15247747-n"/> <!-- wedding day -->
       <SynsetRelation relType="hyponym" target="ewn-15274610-n"/> <!-- day of remembrance, anniversary -->
-      <SynsetRelation relType="hyponym" target="ewn-90001211-a"/>
+      <SynsetRelation relType="hyponym" target="ewn-90001211-n"/>
       <Example>&quot;Mother&apos;s Day&quot;</Example>
     </Synset>
     <!-- Admission Day -->
@@ -14985,7 +14985,7 @@
       <Definition>both morning and night, used when greeting someone</Definition>
       <SynsetRelation relType="hypernym" target="ewn-15137796-n"/> <!-- time period, period, period of time -->
     </Synset>
-    <Synset id="ewn-90001211-a" ili="in" partOfSpeech="a" dc:subject="noun.time" dc:source="Colloquial WordNet">
+    <Synset id="ewn-90001211-n" ili="in" partOfSpeech="n" dc:subject="noun.time" dc:source="Colloquial WordNet">
       <Definition>The day before Valentine&apos;s Day, e.g., February 1</Definition>
       <SynsetRelation relType="hypernym" target="ewn-15182185-n"/> <!-- day -->
     </Synset>
