+++ src/wn-verb.communication.xml
@@ -20209,6 +20209,11 @@
       <SyntacticBehaviour subcategorizationFrame="Somebody ----s something" senses="ewn-tweet-v-90000001-01"/>
       <SyntacticBehaviour subcategorizationFrame="Somebody ----s" senses="ewn-tweet-v-90000001-01"/>
     </LexicalEntry>
+    <LexicalEntry id="ewn-talmbout-v">
+      <Lemma writtenForm="talmbout" partOfSpeech="v"/>
+      <Sense id="ewn-talmbout-v-90006511-01" n="0" synset="ewn-90006511-v" dc:identifier="talmbout%2:10:01::"/>
+      <SyntacticBehaviour subcategorizationFrame="Somebody ----s" senses="ewn-talmbout-v-90006511-01"/>
+    </LexicalEntry>
     <!-- communicate, intercommunicate -->
     <Synset id="ewn-00742582-v" ili="i25403" partOfSpeech="v" dc:subject="verb.communication">
       <Definition>transmit thoughts or feelings</Definition>
@@ -31161,5 +31166,11 @@
       <SynsetRelation relType="hypernym" target="ewn-00995046-v"/> <!-- write -->
       <SynsetRelation relType="hypernym" target="ewn-01073654-v"/> <!-- message -->
     </Synset>
+    <Synset id="ewn-90006511-v" ili="in" partOfSpeech="v" dc:subject="noun.communication" dc:source="Colloquial WordNet">
+      <Definition>to talk about</Definition>
+      <SynsetRelation relType="domain_region" target="ewn-06960420-n"/> <!-- Ebonics, African American English, African American Vernacular English, Black English Vernacular, Black Vernacular, Black English, Black Vernacular English, AAVE -->
+      <SynsetRelation relType="hypernym" target="ewn-00944022-v"/> <!-- talking, talk -->
+    </Synset>
+
   </Lexicon>
 </LexicalResource>
