+++ src/wn-verb.emotion.xml
@@ -5733,7 +5733,6 @@
     <Synset id="ewn-01780873-v" ili="i30620" partOfSpeech="v" dc:subject="verb.emotion">
       <Definition>be fond of</Definition>
       <SynsetRelation relType="is_caused_by" target="ewn-01819595-v"/> <!-- delight, please -->
-      <SynsetRelation relType="hyponym" target="ewn-90003391-v"/>
       <Example>&quot;I like my nephews&quot;</Example>
     </Synset>
     <!-- cotton -->
@@ -5753,7 +5752,6 @@
       <SynsetRelation relType="is_entailed_by" target="ewn-01780389-v"/> <!-- fancy, go for, take to -->
       <SynsetRelation relType="is_caused_by" target="ewn-01819595-v"/> <!-- delight, please -->
       <SynsetRelation relType="is_entailed_by" target="ewn-01824240-v"/> <!-- enjoy, savour, savor, relish, bask -->
-      <SynsetRelation relType="hyponym" target="ewn-90005851-n"/>
       <Example>&quot;I like jogging&quot;</Example>
       <Example>&quot;She likes to read Russian novels&quot;</Example>
     </Synset>
